% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_location.R
\name{check_location}
\alias{check_location}
\title{check_location}
\usage{
check_location(dados, codigos, code, i, variables)
}
\arguments{
\item{dados}{data set containing variable}

\item{codigos}{vector containing codes of states}

\item{code}{column from the data set code e.g. 'cod_regint' or 'cod_mun'}

\item{i}{data set name}

\item{variables}{names of variables from data set}
}
\value{
data set with results
}
\description{
Auxiliary function
}
