% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_report.R
\name{creating_report}
\alias{creating_report}
\title{Creating_report}
\usage{
creating_report(
  mydir = getwd(),
  data_path = NULL,
  fator_mediana = 0.2,
  regras = NULL
)
}
\arguments{
\item{mydir}{Output directory}

\item{path_data}{Path to data}
}
\value{
Report
}
\description{
This function return a report with all statistical analysis and critical rules
}
