% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(caminho, aba = 1, sep = ",", encoding = "UTF-8")
}
\arguments{
\item{caminho}{File path to be uploaded.}

\item{aba}{Sheet name or sheet number (only for .xlsx). Default is 1.}

\item{sep}{Column delimited (only .csv and .txt). Default is ",". For .txt files, use \verb{\\t} if the file is tab-delimited.}

\item{encoding}{Encoding of the .csv or .txt file. Default is "UTF-8".}
}
\value{
A data.frame (or tibble) with the loaded data
}
\description{
Loading data in .xlsx, .csv, .rds or txt.
}
\examples{
dados <- carregar_dados("meus_dados.csv")
dados <- carregar_dados("planilha.xlsx", aba = "Plan2")
dados <- carregar_dados("objeto.rds")
dados <- carregar_dados("dados.txt", sep = "\t")
}
