% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_filter.R
\name{outlier_filter}
\alias{outlier_filter}
\title{Outlier data filter}
\usage{
outlier_filter(variable)
}
\arguments{
\item{variable}{Quantitative variable or indicator}
}
\value{
result
}
\description{
Identifies outliers calculating the upper and lower limits based on three standard desviations.
}
