% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creating_report.R
\name{creating_report}
\alias{creating_report}
\title{Creating_report}
\usage{
creating_report(mydir = getwd(), knitmydir = getwd())
}
\arguments{
\item{mydir}{Output directory}

\item{knitmydir}{output knit}

\item{path_data}{Path to data}
}
\value{
Report
}
\description{
This function return a report with all statistical analysis and critical rules
}
