% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_function.R
\name{outlier_function}
\alias{outlier_function}
\title{Detect outlier}
\usage{
outlier_function(variable)
}
\arguments{
\item{variable}{quantitative variable}
}
\value{
Vector with TRUE or FALSE
}
\description{
Detect outlier in historical data. Here, the boxplot limits are used.
}
